# src/main/resources/api.yaml
openapi: 3.0.0
info:
  title: Sample API
  version: '1.0.0'
  contact:
    email: aserroukh@atsistemas
servers:
  - url: "http://localhost:8000/api"
    description: "local host"
paths:

  /transaction/receive:
    post:
      summary: create new Transaction
      operationId: createTransaction
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: "#/components/schemas/Transaction"
      responses:
        200:
          description: succes
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/StatusTransactionResponse"
        401:
          description: invalid
          content:
            application/json:
              schema:
                type: object
                $ref: "#/components/schemas/StatusTransactionResponse"

  /transaction/{account_iban}:
    get:
      summary: "Search transactions by IBAN number of the account"
      operationId: "seachTransactions"
      parameters:
        - name: "account_iban"
          in: "path"
          description: "IBAN number of the account"
          required: true
          schema:
            type: "string"
        - name: sort
          in: query
          description: Sort by amount
          schema:
            type: string
            enum: [ASC, DESC]
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'

  /transaction/status:
    get:
      summary: "Status and additional information for a specific transaction"
      operationId: "statusTransation"
      requestBody:
        description: "information for a specific transaction"
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransationInfoRequest'
      responses:
        200:
          description: "Retrieves family members by person id"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusTransactionResponse'


components:
  schemas:
    Transaction:
      type: "object"
      required:
        - "reference"
        - "iban"
        - "date"
        - "amount"
        - "fee"
        - "description"
      properties:
        reference:
          type: "string"
          nullable: true
          example: "12345A"
        iban:
          type: "string"
          example: "ES9820385778983000760236"
        date:
          type: "string"
          format: date-time
          nullable: true
          example: "2019-07-16T16:55:42.000Z"

        amount:
          type: "number"
          example: "193.38"
        fee:
          type: "number"
          nullable: true
          example: "3.18"
        description:
          type: "string"
          nullable: true
          example: "Restaurant payment"

    TransationInfoRequest:
      type: object
      properties:
        reference:
          type: string
          example: '12345A'
        channel:
          type: string
          enum: [CLIENT, ATM, INTERNAL]
          example: 'CLIENT'


    StatusTransactionResponse:
      type: "object"
      required:
        - "reference"
        - "status"
        - "amount"
      properties:
        reference:
          type: "string"
          example: "12345A"
        status:
          type: "string"
          enum: [PENDING, SETTLED, FUTURE, INVALID]
          example: "PENDING"
        amount:
          type: "number"
          example: "193.38"
        fee:
          type: "number"
          example: "3.18"